✨ 상위 상품 조회 API 개발

+++ src/main/java/kr/hhplus/be/server/interfaces/product/ProductResponse.java
@@ -16,7 +16,7 @@ public class ProductResponse{
 
     Long price;
 
-    int stock;
+    Long stock;
 
     public static ProductResponse from(Product product) {
         return ProductResponse.builder()
